requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
branca==0.6.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0
django-cors-headers==3.5.0
django-import-export==3.2.0
django-mathfilters==1.0.0
django-sass-processor==1.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
folium==0.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
leven==1.0.4
libsass==0.22.0
MarkupPy==1.14
MarkupSafe==2.1.3
nose==1.3.7
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
prompt-toolkit==3.0.38
psycopg2==2.8.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
XlsxWriter==3.1.0
xlwt==1.3.0
